Removed call to SetMaxLength
It was crashing for me:

➜ openstv git:(master)>env PYTHONPATH=. python openstv/OpenSTV.py
openstv/OpenSTV.py:893: wxPyDeprecationWarning: Call to deprecated item.
  wx.InitAllImageHandlers()
Traceback (most recent call last):
  File "openstv/OpenSTV.py", line 912, in <module>
    app = App(0)
  File "/usr/lib/python2.7/dist-packages/wx-3.0-gtk2/wx/_core.py", line 8628, in __init__
    self._BootstrapApp()
  File "/usr/lib/python2.7/dist-packages/wx-3.0-gtk2/wx/_core.py", line 8196, in _BootstrapApp
    return _core_.PyApp__BootstrapApp(*args, **kwargs)
  File "openstv/OpenSTV.py", line 902, in OnInit
    self.frame = Frame(None)
  File "openstv/OpenSTV.py", line 184, in __init__
    self.console.SetMaxLength(0)
  File "/usr/lib/python2.7/dist-packages/wx-3.0-gtk2/wx/_core.py", line 13339, in SetMaxLength
    return _core_.TextEntryBase_SetMaxLength(*args, **kwargs)
wx._core.PyAssertionError: C++ assertion "IsSingleLine()" failed at ../src/gtk/textctrl.cpp(847) in GetEditable(): shouldn't be called for multiline

Because it says "Shouldn' be called for Multiline" and the flags for
the TextConsole indicate that it is indeed a Multiline, I feel
like removing the SetMaxLength call is safe.